requirements.txt
langgraph>=0.1.1
langchain>=0.2.5
langchain-community>=0.2.5
langchain-openai>=0.1.7
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
streamlit>=1.35.0
requests>=2.31.0
tavily-python>=0.3.3
python-dotenv>=1.0.1
typing-extensions>=4.11.0 # Ensure compatibility for TypedDict Æ